fixed mtl bug.  also changed resolve_model_path() org
